Refactor outdated notification to UI layer
