+++ packages/editor/src/models/collection.ts
@@ -1,6 +1,5 @@
-import React from 'react';
 import { Slicer, SelectorCreator, Parameterizer } from '@rematch/select';
-import { message, notification, Button, Icon } from 'antd';
+import { message } from 'antd';
 import { Node } from 'editure';
 import * as F from 'editure-constants';
 import pick from 'lodash.pick';
@@ -14,9 +13,6 @@ import {
   flattenSteps as flatten,
   unflattenSteps as unflatten,
 } from '@tuture/core';
-
-/** @jsx jsx */
-import { css, jsx } from '@emotion/core';
 
 import {
   FILE,
@@ -338,77 +334,10 @@ export const collection = {
       message.success('保存成功');
       dispatch.drawer.setVisible(false);
     },
-    async openOutdatedNotification() {
-      const key: string = `open${Date.now()}`;
-      const btn = (
-        <Button
-          type="primary"
-          size="small"
-          onClick={() => notification.close(key)}
-        >
-          去处理
-        </Button>
-      );
-      const description = (
-        <div>
-          <p>在你的步骤列表里面存在过时的步骤，请及时迁移内容或删除这些步骤</p>
-          <p
-            css={css`
-              display: flex;
-              flex-direction: row;
-              align-items: center;
-              margin-top: 16px;
-            `}
-          >
-            <Icon
-              type="question-circle"
-              style={{ color: '#096dd9' }}
-              css={css`
-                &:hover {
-                  cursor: pointer;
-                  color: #02b875;
-                }
-              `}
-            />
-            <a
-              href="https://docs.tuture.co/reference/tuture-yml-spec.html#outdated"
-              target="_blank"
-              rel="noopener noreferrer"
-              css={css`
-                margin-left: 4px;
-              `}
-            >
-              什么是过时步骤？
-            </a>
-          </p>
-        </div>
-      );
-      notification.warning({
-        message: '过时步骤提醒',
-        description,
-        btn,
-        key,
-        duration: null,
-        onClick: () => {
-          dispatch.collection.setOutdatedNotificationClicked(true);
-        },
-      });
-    },
     async fetchCollection() {
       try {
         const response = await fetch('/collection');
         const data: Collection = await response.json();
-
-        let outdatedExisted = false;
-
-        if (data?.steps && Array.isArray(data?.steps)) {
-          const len = data.steps.filter((step) => step.outdated).length;
-          outdatedExisted = !!len;
-        }
-
-        if (outdatedExisted) {
-          dispatch.collection.openOutdatedNotification();
-        }
 
         dispatch.collection.setCollectionData(data);
